requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pyTelegramBotAPI==4.10.0
requests==2.28.2
schedule==1.1.0
urllib3==1.26.14
validators==0.20.0
yarl==1.8.2
